Set source and target to 1.8 if it's below 1.8
- Log a warning when this is used

The main reason I'm making this change is that many of the tests
for AST conversion target 1.4 .
By using 1.8 instead of failing in these cases, we can run these test cases properly.
Many of these test cases reveal bugs in our AST converter that we
need to address that are unrelated to source or target level.

I think it's worthwhile to increase the source level,
because this means the user is able to use IDE features before reconfiguring the
project to use Java 8.
However, if people think it's a bad idea to enable this in the IDE,
then I can use an environment variable to enable this,
so that it's only enabled when running the AST conversion tests.

Signed-off-by: David Thompson <davthomp@redhat.com>