Implement completion for field ref in Javadoc
Note that there are two code paths: one through `MemberRef` and
another through an `@link` or `@see` `TagElement`.
This is because the AST conversion tests expect the following to be a
`TextElement` instead of a `MemberRef`:

```java
/**
 * @see MyClass#
 */
```

(and also:)

```java
/**
 * @see #
 */
```

However, the completion tests expect completion for MyClass' members to
work in this case.

Fixes #1103

Signed-off-by: David Thompson <davthomp@redhat.com>