Do not suggest variables from current line for completion
This PR causes a regression: `CompletionTests10.test0007_not_inside_expression`.
I disagree with the results of this test under ECJ;
a type reference to `String` is suggested,
which is why there are 14 completion items.
I don't understand why `String` is suggested.
There used to be 14 completion items because we were suggesting `a`.

Signed-off-by: David Thompson <davthomp@redhat.com>