Improve completion of nested types in Javadoc
This case now works the same as it does upstream:

```java
package dev.datho7561.utils;
/**
 * Outer.Inn|
 */
public class Main {

}
class Outer {
  class Inner {
  }
}
```

There is a bug in jdt.ui which makes it not work in practice though.
It tries to import `Inner`, despite the fact that it's in the same
package and no "additional completion" is provided.

Signed-off-by: David Thompson <davthomp@redhat.com>